pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kor-legal-mcp"
version = "1.0.0"
description = "한국 법령·판례·행정해석 조회 범용 MCP 서버 (law.go.kr Open API 기반)"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0",
    "uvicorn>=0.30.0",
    "starlette>=0.38.0",
    "lxml>=5.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "ruff>=0.5",
]

[project.scripts]
kor-legal-mcp-stdio = "kor_legal_mcp.scripts.run_stdio:main"

[tool.hatch.build.targets.wheel]
packages = ["src/kor_legal_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: requires external APIs (law.go.kr)",
]

[tool.ruff]
line-length = 100
target-version = "py312"
